refactor: split domain and infra layer
