+++ src/domain/user.interface.ts
