+++ src/domain/user.repository.ts
@@ -1,0 +1,8 @@
+import { User, UserID } from './user.interface';
+
+export abstract class UserRepository {
+  abstract list(): IterableIterator<User>;
+  abstract store(user: User): void;
+  abstract load(id: UserID): User | undefined;
+  abstract delete(id: UserID): boolean;
+}
